+++ packages/core/eg/file.ts
@@ -1,8 +1,24 @@
 import { dev } from '@qqi/dev';
-import { readFileToJson } from '../src/index';
+import {
+  getPackageJson,
+  getPackageJsonSync,
+  readFileToJson,
+} from '../src/index';
 
-dev.skip('测试异步读取 JSON 文件', async () => {
-  const a = await readFileToJson('package.json');
+dev('测试文件相关的方法', async test => {
+  test.skip('测试异步读取 JSON 文件', async () => {
+    const a = await readFileToJson('package.json');
+    console.log(a);
+  });
 
-  console.log(a);
+  test.skip('测试同步获取 package 文件', () => {
+    const jsonContent = getPackageJsonSync(5);
+
+    console.log(jsonContent);
+  });
+  test('测试异步获取 package 文件', async () => {
+    const jsonContent = await getPackageJson(4);
+
+    console.log(jsonContent);
+  });
 });