+++ packages/core/eg/runOtherCode.ts
@@ -6,7 +6,7 @@ import { dev } from '@qqi/dev';
 import { waitingTips } from '../src/waiting';
 import { sleep } from 'a-js-tools';
 
-await dev('测试与 run other code 相关的逻辑', async it => {
+await dev.skip('测试与 run other code 相关的逻辑', async it => {
   await it('简单执行 ls ', async () => {
     const result = await runOtherCode({
       code: 'ls',